refactor: unify StatCard and QuickActionCard into DashboardCard
Closes #36
