+++ components/dashboard/stat-card.tsx
@@ -3,26 +3,55 @@ import type { LucideIcon } from "lucide-react"
 import { Card, CardContent } from "@/components/ui/card"
 
 interface StatCardProps {
+  variant?: "stat"
   icon: LucideIcon
   value: number
   label: string
   href: string
 }
 
-export function StatCard({ icon: Icon, value, label, href }: StatCardProps) {
+interface ActionCardProps {
+  variant: "action"
+  icon: LucideIcon
+  title: string
+  description: string
+  href: string
+}
+
+type DashboardCardProps = StatCardProps | ActionCardProps
+
+export function DashboardCard(props: DashboardCardProps) {
+  const { icon: Icon, href } = props
+  const isAction = props.variant === "action"
+
   return (
     <Link href={href}>
       <Card className="transition-shadow hover:shadow-sm">
         <CardContent className="flex items-center gap-4 p-6">
-          <div className="bg-muted flex h-10 w-10 shrink-0 items-center justify-center rounded-lg">
-            <Icon aria-hidden="true" className="text-muted-foreground h-5 w-5" />
+          <div className={`flex h-10 w-10 shrink-0 items-center justify-center rounded-lg ${isAction ? "bg-primary/10" : "bg-muted"}`}>
+            <Icon aria-hidden="true" className={`h-5 w-5 ${isAction ? "text-primary" : "text-muted-foreground"}`} />
           </div>
           <div>
-            <p className="text-2xl font-bold">{value}</p>
-            <p className="text-muted-foreground text-sm">{label}</p>
+            {isAction ? (
+              <>
+                <p className="font-semibold">{props.title}</p>
+                <p className="text-muted-foreground text-sm">{props.description}</p>
+              </>
+            ) : (
+              <>
+                <p className="text-2xl font-bold">{props.value}</p>
+                <p className="text-muted-foreground text-sm">{props.label}</p>
+              </>
+            )}
           </div>
         </CardContent>
       </Card>
     </Link>
   )
 }
+
+// Backward-compatible exports
+export const StatCard = (props: Omit<StatCardProps, "variant">) =>
+  DashboardCard({ ...props, variant: "stat" })
+export const QuickActionCard = (props: Omit<ActionCardProps, "variant">) =>
+  DashboardCard({ ...props, variant: "action" })
